requirements.txt
django>=1.8.0
wheel==0.24.0
# Additional requirements go here
django-mptt==0.7.4
